dashboard + estructura front
