--- src/AccountInfo/AccountInfo.css
@@ -1,135 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter+Tight:wght@300;400;700&display=swap');
-
-.account-container {
-    display: grid;
-    grid-template-columns: 1fr 1fr; /* Ajustar tamaño proporcionalmente */
-    grid-template-rows: auto auto;
-    gap: 20px; /* Aumentar espacio entre recuadros */
-    justify-content: center;
-    align-items: center;
-    height: 90vh; /* Aumentar la altura del fondo */
-    width: 70vw; /* Aumentar el ancho general */
-    margin: auto;
-    background-color: #0d0d0d; /* Hacer el fondo más oscuro */
-    color: white;
-    font-family: 'Inter Tight', sans-serif;
-    padding: 20px; /* Agregar padding para mejor distribución */
-    border-radius: 15px; /* Suavizar los bordes */
-}
-
-.account-box {
-    background-color: #1e1e1e;
-    padding: 15px; /* Ajustar padding */
-    border-radius: 10px;
-    text-align: center;
-    width: 90%; /* Ajustar ancho */
-    height: 260px; /* Aumentar ligeramente la altura uniforme */
-    margin: 0 auto; /* Centrar los recuadros */
-    box-shadow: 0 4px 12px rgba(255, 255, 255, 0.692);
-    display: flex;
-    flex-direction: column;
-    justify-content:  flex-start; /* Alinea el contenido desde arriba */
-    align-items: center; /* Centra el contenido horizontalmente */
-
-}
-
-.account-title {
-    font-size: 16px; /* Ajustar tamaño de fuente */
-    font-weight: bold;
-    margin-bottom: 6px;
-}
-
-.line {
-    border: none;
-    height: 1px;
-    background-color: rgba(255, 255, 255, 0.1);
-    margin-bottom: 10px;
-}
-
-.account-section {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 6px;
-    flex-grow: 1;
-}
-
-.account-button {
-    width: 100%; /* Asegurar que los botones ocupen todo el ancho disponible */
-    max-width: 90%; /* Opcional: evitar que sean demasiado grandes en pantallas grandes */
-    background-color: #79e2ff;
-    color: black;
-    padding: 10px;
-    font-size: 14px;
-    border: none;
-    border-radius: 8px;
-    cursor: pointer;
-    font-weight: bold;
-    text-align: center;
-}
-
-
-
-.account-button:hover {
-    background-color: rgba(121, 226, 255, 0.8);
-}
-
-.account-footer {
-    grid-column: span 2;
-    display: flex;
-    justify-content: center;
-}
-
-.back-button {
-    background-color: #ffffff;
-    color: #000;
-    padding: 6px;
-    font-size: 12px;
-    border: none;
-    border-radius: 8px;
-    cursor: pointer;
-    font-weight: bold;
-    width: 160px;
-    text-align: center;
-}
-
-.back-button:hover {
-    background-color: rgba(255, 255, 255, 0.8);
-}
-
-/* Cambio del título de la tarjeta */
-.account-box:last-child h2 {
-    content: "Método de Pago";
-}
-
-.account-box h2 {
-    font-size: 24px; /* Mantener tamaño uniforme */
-    font-weight: bold;
-    margin-bottom: 16px; /* Aumenta espacio debajo del título */
-    margin-top: 0px; /* Asegura que no tenga margen superior innecesario */
-    text-align: center;
-}
-
-
-.account-option {
-    width: 100%; /* Asegurar que el contenedor del botón se expanda */
-    display: flex;
-    justify-content: center;
-}
-
-.account-box.metodo-pago {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between; /* Mantiene el título arriba y el botón centrado */
-    align-items: center;
-    height: 100%; /* Mantiene el tamaño uniforme */
-    padding-top: 20px; /* Asegura que el título no se mueva */
-}
-
-.account-option.metodo-pago {
-    flex-grow: 1; /* Permite que el botón se mantenga centrado */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-}
