+++ src/App.css
@@ -1,3 +1,4 @@
+
 /* App.css */
 
 /* Reset y box-sizing */
@@ -45,13 +46,4 @@ h1 {
 h2 {
   font-size: 1.2rem;
   color: #ccc;
-}
-
-/* Player al fondo */
-.player-container {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  z-index: 1000;
 }
