--- src/Playlist/Playlist.css
@@ -1,247 +1,0 @@
-
-.layout {
-    width: 1400px;
-    height: 100vh;
-    display: flex;
-    flex-grow: 1;
-    overflow-y: auto;
-    justify-content: flex-end;
-    align-items: center;
-    margin: 0 auto;
-    margin-left: auto;
-    margin-right: 2px;
-}
-
-.box-repro {
-    background: linear-gradient(to bottom, #084d5c, #ffffff);
-    color: white;
-    padding: 40px;
-    border-radius: 12px;
-    width: 1400px;
-    height: 525px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: relative;
-    margin-left: 7%;
-    overflow: hidden;
-}
-.box-sidebar {
-    background: linear-gradient(to bottom, #084d5c, #ffffff);
-    color: white;
-    padding: 20px;
-    border-radius: 12px;
-    width: 300px;
-    height: 875px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: relative;
-    margin-right: 20px;
-}
-.box {
-    background: linear-gradient(to bottom, #5472c9, #121212);
-    color: white;
-    padding: 40px;
-    border-radius: 12px;
-    width: 1400px;
-    height: 875px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: relative;
-    margin-left: 7%;
-    overflow: hidden;
-}
-
-.play-cont {
-    display: flex;
-    width: 100%;
-    max-width: 1200px;
-    height: 300px;
-    align-items: flex-end;
-    justify-content: flex-start;
-    position: absolute;
-    top: 160px;
-    gap: 20px;
-    left: 3%;
-}
-
-
-.image {
-    width: 300px;
-    height: 400px;
-    border-radius: 12px;
-    object-fit: cover;
-}
-
-
-.playlist-info {
-    flex-direction: column;
-    justify-content: flex-end;
-    height: 305px;
-    gap: 10px;
-}
-
-.playlist-info h1 {
-    font-size: 80px;
-    font-weight: bold;
-    color: white;
-    margin: 0;
-}
-
-.playlist-info p {
-    font-size: 18px;
-    color: #b3b3b3;
-    margin: 0;
-}
-
-
-.playlist-actions {
-    display: flex;
-    align-items: center;
-    gap: 15px;
-    position: absolute;
-    top: 330px;
-    left: 50%;
-    transform: translateX(-50%);
-    width: 93%;
-    justify-content: space-between;
-    border-bottom: 2px solid #282828;
-    padding-bottom: 10px;
-}
-
-
-.play-btn {
-    background-color: #1db954;
-    color: white;
-    padding: 15px;
-    border: none;
-    border-radius: 50%;
-    cursor: pointer;
-    font-size: 22px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 55px;
-    height: 55px;
-    transition: background 0.2s;
-}
-
-.play-btn:hover {
-    background-color: #1ed760;
-}
-
-.actions-right {
-    display: flex;
-    align-items: center;
-    gap: 15px;
-    position: absolute;
-    right: 10px;
-}
-
-
-.icon {
-    font-size: 22px;
-    cursor: pointer;
-    color: white;
-    transition: color 0.2s;
-}
-
-.icon:hover {
-    color: #1db954;
-}
-
-
-.song-cont {
-    width: 100%;
-    max-width: 1200px;
-    flex-direction: column;
-    margin-top: 350px;
-    padding-top: 40px;
-    padding-bottom: 20px;
-    overflow-y: auto;
-    height: 100%;
-}
-
-.song-header {
-    display: grid;
-    grid-template-columns: 40px 50px 3fr 2fr 2fr 80px;
-    align-items: center;
-    padding: 10px;
-    border-bottom: 2px solid #282828;
-    font-weight: bold;
-    color: white;
-    font-size: 14px;
-}
-
-.song-list {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    max-width: 1200px;
-}
-
-.song-item {
-    display: grid;
-    grid-template-columns: 40px 50px 3fr 2fr 2fr 80px;
-    align-items: center;
-    padding: 10px;
-    border-bottom: 1px solid #282828;
-    transition: background 0.2s;
-    position: relative;
-}
-
-.song-item:hover {
-    background: rgba(255, 255, 255, 0.1);
-}
-
-
-.song-index {
-    display: block;
-    text-align: center;
-}
-
-.play-icon {
-    display: none;
-    background: #1db954;
-    border: none;
-    border-radius: 50%;
-    font-size: 18px;
-    cursor: pointer;
-    color: white;
-    width: 30px;
-    height: 30px;
-    align-items: center;
-    justify-content: center;
-    transition: background 0.2s;
-}
-
-.song-item:hover .song-index {
-    display: none;
-}
-
-.song-item:hover .play-icon {
-    display: flex;
-}
-
-.song-cover {
-    width: 45px;
-    height: 45px;
-    border-radius: 4px;
-    object-fit: cover;
-}
-
-.song-title {
-    font-weight: bold;
-}
-
-
-.song-date {
-    font-size: 14px;
-    color: #b3b3b3;
-}
-
-.song-duration {
-    text-align: right;
-}
