--- src/Register/Register.jsx
@@ -1,101 +1,0 @@
-import React, { useState } from "react";
-import { useNavigate } from "react-router-dom"; // Importa useNavigate
-import "./Register.css";
-
-function Register() {
-    const [username, setUsername] = useState("");
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
-    const [confirmPassword, setConfirmPassword] = useState("");
-    const navigate = useNavigate(); // Hook para redirigir
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        if (password !== confirmPassword) {
-            alert("Las contraseñas no coinciden");
-            return;
-        }
-        alert(`Registrando usuario: ${username}, Email: ${email}`);
-    };
-
-    const goToLogin = () => {
-        navigate("/login"); // Redirige a la página de inicio de sesión
-    };
-
-    return (
-        <div className="login-container">
-            <div className="login-box">
-                <img
-                    src="../vibrablanco.png"
-                    alt="Vibra Logo"
-                    className="logo"
-                    onClick={() => window.location.reload()}
-                />
-                <h1 className="login-txt">Regístrate en Vibra</h1>
-                <hr className="line"></hr>
-                <form onSubmit={handleSubmit}>
-                    <div className="input-label">
-                        <label htmlFor="username">Nombre de Usuario</label>
-                        <input
-                            type="text"
-                            id="username"
-                            placeholder="Nombre de usuario"
-                            value={username}
-                            onChange={(e) => setUsername(e.target.value)}
-                            required
-                        />
-                    </div>
-
-                    <div className="input-label">
-                        <label htmlFor="email">Correo Electrónico</label>
-                        <input
-                            type="email"
-                            id="email"
-                            placeholder="Correo electrónico"
-                            value={email}
-                            onChange={(e) => setEmail(e.target.value)}
-                            required
-                        />
-                    </div>
-
-                    <div className="input-label">
-                        <label htmlFor="password">Contraseña</label>
-                        <input
-                            type="password"
-                            id="password"
-                            placeholder="********"
-                            value={password}
-                            onChange={(e) => setPassword(e.target.value)}
-                            required
-                        />
-                    </div>
-
-                    <div className="input-label">
-                        <label htmlFor="confirmPassword">Confirmar Contraseña</label>
-                        <input
-                            type="password"
-                            id="confirmPassword"
-                            placeholder="********"
-                            value={confirmPassword}
-                            onChange={(e) => setConfirmPassword(e.target.value)}
-                            required
-                        />
-                    </div>
-
-                    <button type="submit" className="btn-blue">Registrarse</button>
-                </form>
-
-                <div className="divider">o</div>
-
-                <p className="footer-text">
-                    ¿Ya tienes una cuenta?
-                    <span className="login-link" onClick={goToLogin}>
-                        Inicia sesión en Vibra.
-                    </span>
-                </p>
-            </div>
-        </div>
-    );
-}
-
-export default Register;
