--- src/Reproductor/Player.css
@@ -1,84 +1,0 @@
-.player-container {
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    background-color: #282828;
-    color: #fff;
-    padding: 1em;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    box-shadow: 0 -2px 10px rgba(0, 0, 0, 0.5);
-    z-index: 1000;
-}
-
-.musicCover {
-    width: 80px;
-    height: 80px;
-    border-radius: 8px;
-    object-fit: cover;
-}
-
-.info {
-    text-align: center;
-    margin: 0.5em 0;
-}
-
-.title {
-    margin: 0;
-    font-size: 1.2em;
-}
-
-.artist {
-    margin: 0;
-    font-size: 0.9em;
-    color: #b3b3b3;
-}
-
-.controls {
-    display: flex;
-    align-items: center;
-    margin: 0.5em 0;
-}
-
-.controlButton {
-    background: none;
-    border: none;
-    cursor: pointer;
-    margin: 0 1em;
-}
-
-.timelineContainer {
-    width: 90%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.time {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    font-size: 0.8em;
-    color: #b3b3b3;
-    margin-bottom: 0.3em;
-}
-
-.timeline {
-    width: 100%;
-    appearance: none;
-    height: 4px;
-    background-color: #27ae60;
-    border-radius: 2px;
-    outline: none;
-}
-
-.timeline::-webkit-slider-thumb {
-    appearance: none;
-    width: 16px;
-    height: 16px;
-    background: #fff;
-    border-radius: 50%;
-    cursor: pointer;
-    margin-top: -6px;
-}
