--- src/SongItem/SongItem.css
@@ -1,36 +1,0 @@
-.song-item {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 10px;
-    border-bottom: 1px solid #333;
-    transition: background 0.3s;
-}
-
-.song-item:hover {
-    background: #282828;
-}
-
-.song-details {
-    flex-grow: 1;
-    margin-left: 10px;
-}
-
-.song-title {
-    font-size: 16px;
-    font-weight: bold;
-}
-
-.song-artist {
-    font-size: 14px;
-    color: gray;
-}
-
-.song-duration {
-    color: gray;
-}
-
-.play-icon {
-    cursor: pointer;
-    color: #1db954;
-}
