+++ src/components/Player/PlayerStyles.module.css
@@ -1,0 +1,360 @@
+
+/* PlayerStyles.module.css */
+
+/* Variables globales */
+:root {
+    --colorWhite: #ffffff;
+    --colorPrimaryLight1: #ddd;
+    --colorPrimaryLight2: #aaa;
+    --borderRadius: 8px;
+}
+
+/* Contenedor principal del reproductor */
+.playerContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center; /* Centra horizontalmente */
+    justify-content: center; /* Centra verticalmente */
+    width: 100%;
+    margin: auto;
+}
+
+/* Controles (prev, play, next) */
+.controls {
+    margin-top: 2rem;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    gap: 1rem; /* Espacio horizontal entre botones */
+}
+
+/* Botones para prev y next */
+.controlButton {
+    padding: 10px;
+    font-size: 16px;
+    font-weight: bold;
+    border: none;
+    border-radius: 25px;
+    cursor: pointer;
+    color: black;
+    margin-top: 20px;
+    margin-bottom: 5px;
+    background: transparent;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.controlButton:hover {
+    background-color: rgba(39, 174, 96, 0.1);
+}
+
+/* Botón circular grande (Play/Pause) */
+.playerControlPlay {
+    background-color: black; /* Círculo externo */
+    position: relative;
+    width: 10rem; /* Alto y ancho iguales para círculo perfecto */
+    height: 10rem;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 5;
+    overflow: hidden; /* Oculta desbordamientos */
+}
+
+.playerControlPlay::before {
+    content: '';
+    position: absolute;
+    background-color: #1DB954; /* Círculo interno (verde Spotify) */
+    width: 60%;
+    height: 60%;
+    border-radius: 50%;
+    z-index: -2;
+}
+
+/* Contenedor de la imagen de la canción */
+.playerImageContainer {
+    width: 100%;
+    max-width: 300px;
+    height: auto;
+    margin: 0 auto 2rem; /* Centrado y margen inferior */
+    border-radius: var(--borderRadius);
+    overflow: hidden;
+}
+
+.playerImageContainerImg {
+    width: 100%;
+    height: auto;
+    object-fit: cover;
+}
+
+/* Detalles de la canción */
+.playerDetails {
+    text-transform: capitalize;
+}
+
+.playerSubtitle {
+    color: var(--colorPrimaryLight2);
+}
+
+/* Tiempos (por ejemplo, 0:00 / 4:12) */
+.playerTime {
+    margin-top: 2.5rem;
+    display: flex;
+    justify-content: space-between;
+    color: var(--colorPrimaryLight2);
+    font-size: 1.2rem;
+}
+
+/* Íconos generales */
+.playerIcon {
+    transition: all 0.5s;
+}
+
+.playerIconPause,
+.playerIconPlay {
+    width: 4rem;
+    height: 4rem;
+    color: rgba(255, 255, 255, 0.9);
+}
+.time {
+    /* Ocupa todo el ancho disponible */
+    width: 100%;
+    /* Ajusta según tu gusto (80%, 100%, etc.) */
+
+    display: flex;
+    justify-content: space-between; /* "Transcurrido" a la izquierda, "Total" a la derecha */
+    align-items: center;
+    margin-bottom: 1rem; /* Separación con la barra */
+    text-align: center;
+}
+
+.timeSection {
+    display: flex;
+    flex-direction: column;
+    align-items: center; /* Si quieres centrar el texto verticalmente */
+    /* Opcional: margin-left, margin-right si deseas más separación */
+}
+
+.timeLabel {
+    font-size: 1rem;
+    color: #ccc; /* Un color gris claro */
+    margin-bottom: 0.2rem; /* Un pequeño espacio entre el label y el valor */
+}
+
+.timeValue {
+    font-size: 1.2rem;
+    font-weight: bold;
+    /* Ajusta el color si quieres resaltarlo */
+    color: #fff;
+}
+/* Animación de rotación (opcional) */
+@keyframes rotate {
+    from {
+        transform: rotate(0);
+    }
+    to {
+        transform: rotate(360deg);
+    }
+}
+
+/* Contenedor para la barra de progreso y tiempos */
+.timelineContainer {
+    margin-top: 1rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 80%; /* Ajusta según el ancho deseado */
+}
+
+/* Barra de progreso (input range) */
+.playerProgress {
+    -webkit-appearance: none;
+    border-radius: 2rem;
+    background-color: #ccc;
+    width: 100%;
+    height: 1rem;
+    margin-top: 1rem;
+    margin-bottom: 1rem;
+    position: relative;
+}
+
+:global(input[type="range"].playerProgress) {
+    /* Anulaciones específicas que tienen mayor prioridad */
+    width: 100%;
+    padding: 0;
+    margin-bottom: 1rem;
+    border: none;
+}
+
+.playerProgress:focus {
+    border: none;
+    outline: none;
+}
+
+/* Estilo para navegadores WebKit */
+.playerProgress::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    appearance: none;
+    width: 2rem;
+    height: 2rem;
+    border-radius: 50%;
+    background: #1DB954;
+    cursor: pointer;
+    margin-top: -0.5rem;
+}
+
+.playerProgress::-webkit-slider-runnable-track {
+    height: 1rem;
+    border-radius: 2rem;
+    background: #ccc;
+}
+
+/* Estilo para navegadores Mozilla */
+.playerProgress::-moz-range-thumb {
+    width: 2rem;
+    height: 2rem;
+    border-radius: 50%;
+    background: #1DB954;
+    cursor: pointer;
+    margin-top: -0.5rem;
+}
+
+.playerProgress::-moz-range-track {
+    height: 1rem;
+    border-radius: 2rem;
+    background: #ccc;
+}
+
+/* Opcional: estilos para el mini reproductor */
+.miniPlayer {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    position: absolute;
+    padding: 1rem;
+    top: 85vh;
+    left: 0;
+    z-index: 2000;
+    background-color: var(--colorPrimaryLight2);
+    /* Por defecto oculto; se activa en móviles si se requiere */
+    display: none;
+}
+
+.miniPlayerMeta {
+    display: flex;
+    align-items: center;
+    width: 35%;
+}
+
+.miniPlayerProgressCont {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 5rem;
+}
+
+@media (max-width: 37.5em) {
+    .miniPlayerProgressCont {
+        padding: 0 2rem;
+    }
+    .miniPlayerProgressCont {
+        padding: 0;
+    }
+}
+
+.miniPlayerProgress {
+    flex: 1;
+    margin: 0 1rem;
+}
+
+.miniPlayerMedia {
+    width: 5rem;
+    height: 5rem;
+    margin: 0 1.5rem;
+    border-radius: 1rem;
+    overflow: hidden;
+}
+
+@media (max-width: 51.25em) {
+    .miniPlayerMedia {
+        width: 4rem;
+        height: 4rem;
+        margin: 0 0.5rem;
+    }
+}
+
+@media (max-width: 35.625em) {
+    .miniPlayerMedia {
+        display: none;
+    }
+}
+
+.miniPlayerMediaImg {
+    width: 100%;
+    height: 100%;
+}
+
+.miniPlayerName {
+    text-transform: capitalize;
+    font-size: 1.5rem;
+}
+
+@media (max-width: 51.25em) {
+    .miniPlayerName {
+        font-size: 1.3rem;
+    }
+}
+
+@media (max-width: 35.625em) {
+    .miniPlayerName {
+        font-size: 1rem;
+    }
+}
+
+.miniPlayerControls {
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+}
+
+.miniPlayerControlsPlay {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.miniPlayerButtons {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.miniPlayerIcon {
+    margin: 0 2rem;
+    width: 2rem;
+    height: 2rem;
+}
+
+@media (max-width: 37.5em) {
+    .miniPlayerIcon {
+        margin: 0 0.5rem;
+    }
+}
+
+.miniPlayerIconPlay,
+.miniPlayerIconPause {
+    width: 3rem;
+    height: 3rem;
+    color: rgba(0, 0, 0, 0.7);
+}
+
+@media (max-width: 51.25em) {
+    .miniPlayerIconPlay,
+    .miniPlayerIconPause {
+        width: 3rem;
+        height: 3rem;
+    }
+}
