+++ src/components/PlaylistGrid/PlaylistGrid.css
@@ -1,0 +1,25 @@
+.playlist-grid {
+    display: flex;
+    justify-content: center;
+    gap: 20px;
+}
+
+.playlist-card {
+    text-align: center;
+    cursor: pointer;
+    width: 150px;
+    background: white;
+    padding: 10px;
+    border-radius: 15px;
+    box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+.playlist-card img {
+    width: 100%;
+    border-radius: 10px;
+    transition: transform 0.3s ease;
+}
+
+.playlist-card img:hover {
+    transform: scale(1.05);
+}
