+++ src/components/ProfileCard/ProfileCard.css
@@ -1,0 +1,45 @@
+.profile-card {
+    display: flex;
+    align-items: center;
+    background: white;
+    padding: 15px;
+    border-radius: 15px;
+    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.1);
+    width: 250px;
+}
+
+.profile-img {
+    width: 60px;
+    height: 60px;
+    border-radius: 50%;
+}
+
+.profile-info {
+    text-align: left;
+    margin-left: 15px;
+}
+
+.profile-info h3 {
+    font-size: 1rem;
+    margin: 0;
+    font-weight: bold;
+    color: #2c3e50; /* Gris oscuro */
+}
+
+.profile-info p {
+    font-size: 0.9rem;
+    color: #7f8c8d;
+    margin: 0;
+}
+
+.logout-btn {
+    background: none;
+    border: none;
+    cursor: pointer;
+    color: black;
+    font-size: 1.2rem;
+}
+
+.logout-btn:hover {
+    color: red;
+}
