+++ src/pages/Dashboard/Dashboard.css
@@ -1,0 +1,70 @@
+.dashboard-container {
+    padding: 20px;
+}
+
+/* 🎵 Secciones con bordes redondeados */
+.section-container {
+    background: white;
+    border-radius: 15px;
+    padding: 20px;
+    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.1);
+    margin-bottom: 20px;
+}
+
+/* 🔍 Barra de búsqueda */
+.search-bar {
+    width: 100%;
+    padding: 10px;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+}
+
+/* 🎵 Playlists */
+.playlist-grid {
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+}
+
+.playlist-card {
+    text-align: center;
+    cursor: pointer;
+    width: 160px;
+    background: white;
+    padding: 12px;
+    border-radius: 15px;
+    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.1);
+}
+
+.playlist-card img {
+    width: 100%;
+    border-radius: 10px;
+    transition: transform 0.3s ease;
+}
+
+.playlist-card img:hover {
+    transform: scale(1.05);
+}
+
+/* 🎶 Lista de canciones más compacta */
+.song-section h3 {
+    font-size: 1.2rem;
+    font-weight: bold;
+}
+
+.song-item {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 10px;
+    border-radius: 10px;
+    transition: background 0.3s ease;
+}
+
+.song-item:hover {
+    background: #f0f0f0;
+}
+
+.song-title {
+    font-size: 1rem;
+}
