+++ src/pages/Home/Home.css
@@ -1,0 +1,51 @@
+.home-container {
+    background: white;
+    border-radius: 20px;
+    padding: 30px;
+    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.1);
+    width: 100%;
+    max-width: 1100px;
+}
+
+/* 🎵 Título más elegante */
+.home-title {
+    font-size: 2rem;
+    font-weight: bold;
+    color: #2c3e50;
+    text-align: center;
+}
+
+/* 📜 Texto más compacto */
+.home-subtitle {
+    font-size: 1rem;
+    color: #7f8c8d;
+    text-align: center;
+    margin-bottom: 20px;
+}
+
+/* 🎶 Playlists bien alineadas */
+.playlist-container {
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+    gap: 20px;
+}
+
+.playlist-card {
+    width: 160px;
+    background: white;
+    padding: 12px;
+    border-radius: 15px;
+    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.1);
+    text-align: center;
+    transition: transform 0.3s ease;
+}
+
+.playlist-card img {
+    width: 100%;
+    border-radius: 10px;
+}
+
+.playlist-card:hover {
+    transform: scale(1.05);
+}
