+++ src/pages/Playlist/Playlist.css
@@ -1,0 +1,153 @@
+
+.layout {
+    display: flex;
+    width: 100%;
+    height: 100vh;         /* Ocupa toda la altura de la ventana */
+    margin: 0;
+    padding: 0;
+    overflow: hidden;       /* Evita que la página crezca y tenga scroll global */
+}
+
+/* Columna izquierda (sidebar) */
+.box-sidebar {
+    width: 25%; /* o el valor que quieras */
+    background: linear-gradient(to bottom, #084d5c, #ffffff);
+    color: white;
+    border-radius: 12px;
+
+    /* Ocupa toda la altura de la pantalla */
+    height: 100vh;
+    box-sizing: border-box;
+
+    /* Si quieres que la barra lateral siga fija al hacer scroll en la página: */
+    position: sticky;
+    top: 0;
+
+    /* Flexbox para centrar en vertical y horizontal */
+    display: flex;
+    flex-direction: column;
+    justify-content: center; /* Centra verticalmente */
+    align-items: center;     /* Centra horizontalmente */
+
+    /* Si la sidebar tiene contenido extra (menús, enlaces, etc.)
+       y deseas que el reproductor se mantenga centrado,
+       pon ese contenido por arriba o abajo y ajusta su estilo en consecuencia.
+       Si el contenido es demasiado largo, la sidebar podría scrollear. */
+    overflow-y: auto;
+    /* Con 'overflow-y: auto;', si la sidebar sobrepasa la pantalla,
+       se activará un scroll interno en la columna. */
+}
+
+
+/* Columna derecha (contenido principal) */
+.box {
+    flex: 1;               /* Toma el resto del espacio horizontal */
+    background: linear-gradient(to bottom, #5472c9, #121212);
+    color: white;
+    padding: 40px;
+    border-radius: 12px;
+
+    /* Para que esta columna también tenga scroll interno y no se extienda la página */
+    height: 100vh;
+    overflow-y: auto;
+    box-sizing: border-box;
+}
+
+/* Encabezado con la portada y el nombre de la playlist */
+.play-cont {
+    display: flex;
+    width: 100%;
+    max-width: 1200px;
+    height: auto;             /* Quita la altura fija */
+    align-items: flex-end;
+    justify-content: flex-start;
+    gap: 20px;
+    margin-bottom: 2rem;      /* En lugar de position absolute */
+}
+
+/* Imagen de la playlist */
+.image {
+    width: 300px;
+    height: 400px;
+    border-radius: 12px;
+    overflow: hidden;
+}
+
+.image img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+/* Información de la playlist (título, etc.) */
+.playlist-info {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    gap: 10px;
+    color: #fff;
+}
+
+.playlist-info h1 {
+    font-size: 80px;
+    font-weight: bold;
+    margin: 0;
+}
+
+.playlist-info p {
+    font-size: 18px;
+    color: #b3b3b3;
+    margin: 0;
+}
+
+/* Acciones de la playlist (botón play, iconos) */
+.playlist-actions {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+    width: 100%;               /* Se extiende al ancho disponible */
+    border-bottom: 2px solid #282828;
+    padding-bottom: 10px;
+    margin-bottom: 2rem;       /* Espacio debajo para separar de la lista */
+    justify-content: space-between;
+    /* Eliminado position absolute, top, left, transform */
+}
+
+/* Botón principal de play */
+.play-btn {
+    background-color: #1db954;
+    color: white;
+    padding: 15px;
+    border: none;
+    border-radius: 50%;
+    cursor: pointer;
+    font-size: 22px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 55px;
+    height: 55px;
+    transition: background 0.2s;
+}
+
+.play-btn:hover {
+    background-color: #1ed760;
+}
+
+/* Iconos de la derecha */
+.actions-right {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+}
+
+.icon {
+    font-size: 22px;
+    cursor: pointer;
+    color: white;
+    transition: color 0.2s;
+}
+
+.icon:hover {
+    color: #1db954;
+}
