+++ src/pages/Subs/Subs.css
@@ -1,0 +1,106 @@
+
+/* Estilos para la sección de suscripciones */
+.subs-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 20px;
+    padding: 20px;
+    flex-wrap: wrap;
+}
+
+.line {
+    border: none;
+    height: 1px;
+    margin: 16px 0px 0px;
+    color: rgba(255, 255, 255, 0.1);
+    background-color: rgba(255, 255, 255, 0.1);
+}
+
+/* Logo */
+.logo {
+    width: 140px;
+    margin-top: 0px;
+    margin-bottom: 0px;
+    cursor: pointer;
+}
+
+.Premium {
+    margin-top: 0px;
+    margin-bottom: 35px;
+    font-size: 45px;
+    font-weight: bold;
+    font-family: 'Inter Tight', sans-serif;
+    color: #b0c4de
+}
+
+.Gratuito {
+    margin-top: 0px;
+    margin-bottom: 35px;
+    font-size: 45px;
+    font-weight: bold;
+    font-family: 'Inter Tight', sans-serif;
+    color: #ffb6c1
+}
+
+/* Estilos para cada tarjeta de suscripción */
+.sub-card {
+    background-color: #1e1e1e;
+    color: white;
+    padding: 20px;
+    border-radius: 10px;
+    text-align: center;
+    width: 300px;
+    border: 1px solid #616060dc;
+    box-shadow: 0 4px 8px rgba(255, 255, 255, 0.692);
+}
+
+.sub-card h3 {
+    font-size: 22px;
+    margin-bottom: 10px;
+}
+
+.price {
+    font-size: 18px;
+    font-weight: bold;
+    margin-bottom: 5px;
+}
+
+.after {
+    font-size: 14px;
+    color: #b3b3b3;
+    margin-bottom: 15px;
+}
+
+ul {
+    list-style-type: none;
+    padding: 0;
+    margin-bottom: 15px;
+}
+
+ul li {
+    margin: 5px 0;
+}
+
+/*button { //Comentado porque afecta a todos los botones
+    width: 100%;
+    padding: 10px;
+    font-size: 16px;
+    font-weight: bold;
+    border: none;
+    border-radius: 25px;
+    cursor: pointer;
+    color: black;
+    margin-top: 20px;
+    margin-bottom: 5px;
+}
+
+button:hover {
+    opacity: 0.8;
+}*/
+
+.note {
+    font-size: 12px;
+    color: #b3b3b3;
+    margin-top: 10px;
+}
